requirement.txt
mt5linux
pyyaml
telethon
asyncio